requirements.txt
firebird-base==1.8.0
firebird-driver==1.10.9
numpy==2.2.4
pandas==2.2.3
protobuf==6.30.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
tzdata==2025.2
